new changes

+++ src/controller/categoryController.js
@@ -11,8 +11,23 @@ module.exports.getAllCategories = (req, res) => {
     )
 }
 
-module.exports.addCategory = (req, res) => {
+module.exports.getSingleCategory = (req, res) => {
   categoryModel
+    .getSingleCategoryData(req.params.id)
+    .then( ([rows, metadata]) => res.status(200).json(JSON.stringify(rows)))
+    .catch((err) =>
+      res.status(400).send({
+        message: err
+      })
+    )
+}
+
+module.exports.addCategory = (req, res, next) => {
+  if(!req.user.admin){
+    res.status(401).json({ message: "sorry you are not admin"});
+  }
+  else {
+    categoryModel
     .createCategory(req.body.name, req.body.details)
     .then(res.status(201).json({ message: "category created" }))
     .catch((err) =>
@@ -20,20 +35,31 @@ module.exports.addCategory = (req, res) => {
         message: err,
       })
     );
+  }
+  
 };
 
 module.exports.updateCategory = (req, res) => {
-  categoryModel
-  .updateCategory(req.params.id, req.body.name, req.body.details)
-  .then(res.status(200).json({ message: "category updated" }))
-  .catch((err) =>
-    res.status(400).send({
-      message: err,
-    })
-  );
+  if(!req.user.admin){
+    res.status(401).json({ message: "sorry you are not admin"});
+  }
+  else {
+    categoryModel
+    .updateCategory(req.params.id, req.body.name, req.body.details)
+    .then(res.status(200).json({ message: "category updated" }))
+    .catch((err) =>
+      res.status(400).send({
+        message: err,
+      })
+    );
+  }
 };
 
 module.exports.deleteCategory = (req, res) => {
+  if(!req.user.admin){
+    res.status(401).json({ message: "sorry you are not authorised"});
+  }
+  else {
     categoryModel
     .deleteCategory(req.params.id)
     .then(res.status(200).json({ message: "category deleted" }))
@@ -42,6 +68,7 @@ module.exports.deleteCategory = (req, res) => {
         message: err,
       })
     );
+  }
 };
